fix callbration scripts

+++ r_scripts/docking/r_scripts/calibration_merge_structure.R
@@ -3,27 +3,34 @@ library(bio3d)
 library(readr)
 library(dplyr)
 library(ggplot2)
-#v_rmsds<-seq(from=0,to=100,by=1)
+v_rmsds<-seq(from=0,to=100,by=1)
 setwd(part_analysis)
 part<-paste0(part_analysis,"din/")
 setwd(part)
 
 df_all<-read.csv(paste0(part_analysis,"df_all.csv"),stringsAsFactors = F)
 df_all<-df_all%>%mutate(name=paste0(receptor,"_",ligand))
-if(dir.exists("df_RMSD_merge")){dir.create("df_RMSD_merge")}
+if(!dir.exists("df_RMSD_merge")){dir.create("df_RMSD_merge")}
 df_all$center<-NULL
 df_all<-unique(df_all)
 for (i in 1:nrow(df_all)) {
   if(!file.exists(paste0("df_RMSD_merge/",df_all$name[i],".csv"))){
-    df_RMSD_all<-read.csv(paste0("RMSD_merged/",df_all$name[i],".csv"),stringsAsFactors = F)
-    df_RMSD_all<-df_RMSD_all%>%mutate(RMSD=round(RMSD,digits = 1))
-    df_RMSD_all<-df_RMSD_all%>%group_by(RMSD)%>%mutate(number=n())
-    df_RMSD_all<-df_RMSD_all%>%select(RMSD,number)
-    df_RMSD_all<-unique(df_RMSD_all)
-    write.csv(df_RMSD_all,paste0("df_RMSD_merge/",df_all$name[i],".csv"),row.names = F)
+    if(file.exists(paste0("RMSD_merged/",df_all$name[i],".csv"))){
+      df_RMSD_all<-read.csv(paste0("RMSD_merged/",df_all$name[i],".csv"),stringsAsFactors = F)
+      df_RMSD_all<-df_RMSD_all%>%mutate(RMSD=round(RMSD,digits = 1))
+      df_RMSD_all<-df_RMSD_all%>%group_by(RMSD)%>%mutate(number=n())
+      df_RMSD_all<-df_RMSD_all%>%select(RMSD,number)
+      df_RMSD_all<-unique(df_RMSD_all)
+      write.csv(df_RMSD_all,paste0("df_RMSD_merge/",df_all$name[i],".csv"),row.names = F)
+    }
   }
 }
-
+for (i in 1:nrow(df_all)) {
+  if(!file.exists(paste0("df_RMSD_merge/",df_all$name[i],".csv"))){
+    df_all$name[i]<-NA
+  }
+}
+df_all<-df_all%>%filter(!is.na(name))
 df_RMSD_all<-read.csv(paste0("df_RMSD_merge/",df_all$name[1],".csv"),stringsAsFactors = F)
 for (i in 2:nrow(df_all)) {
   df_RMSD_add<-read.csv(paste0("df_RMSD_merge/",df_all$name[i],".csv"),stringsAsFactors = F)
